services/user/requirements.txt
click~=8.1.3
click-config-file~=0.6.0
grpcio-tools~=1.49.1
toml~=0.10.2
